llm-analytics/anthropic-python/main.py
import os
from dotenv import load_dotenv
from posthog.ai.anthropic import Anthropic
from posthog import Posthog

load_dotenv()

posthog = Posthog(
    os.environ["POSTHOG_API_KEY"],
    host=os.environ["POSTHOG_HOST"],
)

client = Anthropic(
    api_key=os.environ["ANTHROPIC_API_KEY"],
    posthog_client=posthog,
)

response = client.messages.create(
    model="claude-sonnet-4-20250514",
    max_tokens=1024,
    messages=[
        {"role": "user", "content": "Tell me a fun fact about hedgehogs"}
    ],
    posthog_distinct_id="user_123",
    posthog_trace_id="trace_123",
    posthog_properties={"conversation_id": "abc123"},
)

print(response.content[0].text)
